Correction erreurs compilation

+++ src/headers/spanning_tree_protocol.h
@@ -20,7 +20,7 @@ void receive_bpdu(Device *device, Frame *bpdu);
 void elect_root_bridge(Network *network);
 void visite_composante_connexe(Network *network, Device *device, bool *visite);
 uint32_t nb_composantes_connexes(Network *network);
-void dijkstra(Network *network, Device device, double *distance_sommet);
+void dijkstra(Network *network, Device device, uint16_t *distance_sommet);
 Device * find_shortest_path(Network * network, Device *device, Device *destination, Device * path[]);
 void update_ports_state(Device *device, Device *destination, Device * path[]);
 // uint8_t compare_bpdu(BPDU *bpdu1, BPDU *bpdu2);
